wip gpg store
