+++ schema.json
@@ -9,13 +9,9 @@
       "type": "string",
       "description": "Name of the service. parameters will be prefixed by the value provided"
     },
-    "db_dir": {
-      "type": "string",
-      "description": "Path to directory where the database should be stored. \nEg: ~/my-path. The database will be in ~/my-path/<service>. \nYou can use relative or absolute path"
-    },
     "provider": {
       "type": "string",
-      "enum": ["ssm", "secrets-manager", "gpg"],
+      "enum": ["ssm", "secrets-manager"],
       "default": "ssm",
       "description": "Deploy parameters to the given provider. Eg. ssm, secrets-manager"
     },
